requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
billiard==4.2.1
celery==5.5.3
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==45.0.4
deprecated==1.2.18
deprecation==2.1.0
distro==1.9.0
dnspython==2.7.0
easyocr==1.7.2
email-validator==2.2.0
fastapi==0.115.12
fastapi-cli==0.0.7
filelock==3.18.0
fsspec==2025.5.1
h11==0.16.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
idna==3.10
imageio==2.37.0
img2pdf==0.6.1
jinja2==3.1.6
jiter==0.10.0
kombu==5.5.4
lazy-loader==0.4
lxml==5.4.0
markdown-it-py==3.0.0
markupsafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
networkx==3.5
ninja==1.11.1.4
numpy==2.3.0
ocrmypdf==16.10.2
openai==1.87.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
packaging==25.0
pdf2image==1.17.0
pdfminer-six==20250506
pi-heif==0.22.0
pikepdf==9.8.1
pillow==11.2.1
pluggy==1.6.0
prompt-toolkit==3.0.51
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.7
pydantic-core==2.33.2
pygments==2.19.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pyyaml==6.0.2
redis==5.2.1
rich==14.0.0
rich-toolkit==0.14.7
scikit-image==0.25.2
scipy==1.15.3
setuptools==80.9.0
shapely==2.1.1
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.6.11
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typer==0.16.0
typing-extensions==4.14.0
typing-inspection==0.4.1
tzdata==2025.2
uvicorn==0.34.3
vine==5.1.0
watchfiles==1.1.0
wcwidth==0.2.13
websockets==15.0.1
wrapt==1.17.2
